% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeRMSE.R
\name{computeRMSE}
\alias{computeRMSE}
\title{compute the Root Mean Squared Error for parameters in A and S}
\usage{
computeRMSE(A_est, S_est, Apop, Spop, Afree, Sfree)
}
\description{
compute the Root Mean Squared Error for parameters in A and S
}
\examples{

}
\author{
Jannik Orzek
}
