% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createARCLModel.R
\name{createARCLModel}
\alias{createARCLModel}
\title{createARCLModel}
\usage{
createARCLModel(sampleSize, numLatent, Timepoints, burning, Avalues, Afree,
  Alabel, Svalues, Sfree, Slabel, S_firstObsAllFree = TRUE,
  SimulatedDataSet)
}
\arguments{
\item{sampleSize}{sample size}

\item{numLatent}{number of latent varaibles per time point}

\item{Timepoints}{number of time points observed in the final data set}

\item{burning}{number of observations to burn before the initial time point}

\item{Avalues}{values in ARCL matrix}

\item{Afree}{free parameters in ARCL matrix}

\item{Alabel}{labels parameters for ARCL matrix}

\item{Svalues}{values in S matrix}

\item{Sfree}{free parameters in S matrix}

\item{Slabel}{labels parameters for S matrix}

\item{S_firstObsAllFree}{should all initial observations be allowed to correlate?}
}
\description{
createARCLModel creates an ARCL(1)-SEM model with OpenMx with >= 1 latent variables and 1 manifest per latent and time point for data simulation. It returns the covariance based and a raw data based model
}
\examples{

library(OpenMx)

Alabel <- matrix(c("a11", "a12", "a13", "a14", "a15", "a21", "a22", "a23", "a24", "a25", "a31", "a32", "a33", "a34", "a35", "a41", "a42", "a43", "a44", "a45", "a51", "a52", "a53", "a54", "a55"), nrow = 5, byrow = T)

Avalues <- diag(.5,nrow=5,ncol = 5)
Avalues[2:5,1] <-.3
Avalues[3:5,2] <-.3
Avalues[4,5] <-.3
Afree <- matrix(TRUE, 5,5)
Alabel <- matrix(c("a11", "a12", "a13", "a14", "a15", "a21", "a22", "a23", "a24", "a25", "a31", "a32", "a33", "a34", "a35", "a41", "a42", "a43", "a44", "a45", "a51", "a52", "a53", "a54", "a55"), nrow = 5, byrow = T)

Svalues <- diag(.75,nrow=5,ncol = 5)
Sfree <- diag(TRUE, 5,5)
Slabel <- matrix(c("s11", NA, NA, NA, NA,
                   NA, "s22", NA, NA, NA,
                   NA, NA, "s33", NA, NA,
                   NA, NA, NA, "s44", NA,
                   NA, NA, NA, NA, "s55"), nrow = 5, byrow = T)


Svalues_init <- diag(1,nrow=5,ncol = 5)
Sfree_init <- diag(TRUE, 5,5)
Slabel_init <- matrix(NA, nrow = 5, byrow = T)

simObj <- simARCL(numLatent = 5, Timepoints = 4, burning = 10,
                  Avalues = Avalues, Alabel = Alabel, Afree = Afree,
                  Svalues = Svalues, Sfree = Sfree, Slabel = Slabel,
                  Svalues_init = Svalues_init, Sfree_init = Sfree_init, Slabel_init = Slabel_init,
                  sampleSize = 1000, S_firstObsAllFree = T)


temp <- mxRun(simObj$mxARCL_cov)
summary(temp)

temp2 <- mxTryHard(simObj$mxARCL_FIML)
summary(temp2)

}
\author{
Jannik Orzek
}
