% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateARCL_1_mx.R
\name{generateARCL_1_mx}
\alias{generateARCL_1_mx}
\title{generateARCL_1_mx}
\usage{
generateARCL_1_mx(timepoints, burning, sample_size, raw_data,
  lastLVsCorrelated = FALSE)
}
\arguments{
\item{timepoints}{number of time points observed in the final data set}

\item{burning}{number of observations to burn before the initial time point}

\item{sample_size}{sample size}

\item{raw_data}{raw data set - e.g. simulated with generateARCL_1_lavaan}

\item{lastLVsCorrelated}{should the last latent variables be correlated?}
}
\description{
generateARCL_1_mx creates an ARCL(1)-SEM model syntax for OpenMx with 2 latent variables and 1 manifest per latent and time point for data simulation
Important: the naming of the variables starts at t = 0 and increases to t = timepoints+burning-1. The data-points used in the analysis
are: t_burning (corresponds to the first observation in the data set because the indexing starts at 0) to t_(burning+timepoints-1).
}
\examples{
library(lavaan)
library(OpenMx)
a_11 = .5
a_22 = .5
a_12 = .5
a_21 = 0
phi1_init = 1
phi2_init = 1
timepoints = 10
burning = 100
sample_size = 100
seed = 1234
lastLVsCorrelated = FALSE

# generate lavaan ARCL(1) with two latent variables:
lavaanModel <- generateARCL_1(a_11 = a_11, a_22 = a_22,a_12 = a_12,a_21 = a_21,
                             phi1_init = phi1_init, phi2_init = phi2_init,
                             timepoints = timepoints, burning = burning,
                             sample_size = sample_size, seed = seed, lastLVsCorrelated = lastLVsCorrelated)

myModel <- cfa(lavaanModel$AnalysisModel, sample.cov = cov(lavaanModel$raw_data[,(2*burning+1):(2*burning+2*timepoints)]), sample.nobs = lavaanModel$sample_size)
summary(myModel)

mymxModel <- generateARCL_1_mx(timepoints = timepoints, burning = burning, sample_size = sample_size, raw_data = lavaanModel$raw_data)

fitMxARCL <- mxTryHard(mymxModel)
summary(fitMxARCL)

}
\author{
Jannik Orzek
}
