% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQdeltaT.R
\name{getQdeltaT}
\alias{getQdeltaT}
\title{getQdeltaT}
\usage{
getQdeltaT(DRIFT, DIFFUSION, deltaT)
}
\arguments{
\item{DIFFUSION}{lower triangle diffusion matrix}
}
\description{
getQdeltaT
}
\examples{

}
\author{
Jannik Orzek
}
