% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUniqueA.R
\name{getUniqueA}
\alias{getUniqueA}
\title{getUniqueA}
\usage{
getUniqueA(ARCL_AMatrix)
}
\arguments{
\item{ARCL_AMatrix}{A matrix from mxModel (full mxMatrix, not just values)}
}
\description{
Extracts the unique elements from an mxModel A-matrix. Used for ARCL models with equality constraints
}
\examples{

}
\author{
Jannik Orzek
}
