% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simARCL_scaled_20.R
\name{simARCL_scaled_20}
\alias{simARCL_scaled_20}
\title{simARCL_scaled_20}
\usage{
simARCL_scaled_20(sampleSize, autoEffect, crossEffect)
}
\description{
simARCL_scaled_20 creates an ARCL(1)-SEM model with OpenMx with 5 latent variables and 1 manifest per latent and time point for data simulation. It returns the simulated data set and a covariance based and a raw data based model. 10 Timepoints are created of which the initial five are omitted (burned)
The model has 20% non-zero cross-lagged effects
}
\author{
Jannik Orzek
}
