% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation1.R
\name{simulation1}
\alias{simulation1}
\title{Simulation study 1}
\usage{
simulation1(sampleSize, seed, wd)
}
\arguments{
\item{sampleSize}{sampleSize}

\item{seed}{seed}

\item{wd}{working directory. The results will be saved here}
}
\description{
Simulation study 1
}
\examples{

}
\author{
Jannik Orzek
}
