% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation2.R
\name{simulation2}
\alias{simulation2}
\title{Simulation study 2}
\usage{
simulation2(sampleSize, seed, wd, gamma, total_repetitions)
}
\arguments{
\item{sampleSize}{sampleSize}

\item{seed}{seed}

\item{wd}{working directory. The results will be saved here}

\item{gamma}{gamma}

\item{total_repetitions}{total number of repetitions}
}
\description{
Simulation study 2
}
\examples{

}
\author{
Jannik Orzek
}
