% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation3_15.R
\name{simulation3_15}
\alias{simulation3_15}
\title{Simulation study 3
15 percent of cross-lagged parameters non-zero}
\usage{
simulation3_15(sampleSize, seed, wd, total_repetitions, crossEffect)
}
\arguments{
\item{sampleSize}{sampleSize}

\item{seed}{seed}

\item{wd}{working directory. The results will be saved here}

\item{total_repetitions}{total number of repetitions}

\item{crossEffect}{size of cross-lagged effects}
}
\description{
Simulation study 3
15 percent of cross-lagged parameters non-zero
}
\examples{

}
\author{
Jannik Orzek
}
