% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation3_30.R
\name{simulation3_30}
\alias{simulation3_30}
\title{Simulation study 3
30 percent of cross-lagged parameters non-zero}
\usage{
simulation3_30(sampleSize, seed, wd, total_repetitions)
}
\arguments{
\item{sampleSize}{sampleSize}

\item{seed}{seed}

\item{wd}{working directory. The results will be saved here}

\item{total_repetitions}{total number of repetitions}
}
\description{
Simulation study 3
30 percent of cross-lagged parameters non-zero
}
\examples{

}
\author{
Jannik Orzek
}
