% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation4.R
\name{simulation4}
\alias{simulation4}
\title{Simulation study 4}
\usage{
simulation4(sampleSize, seed, wd, total_repetitions, Scale)
}
\arguments{
\item{sampleSize}{sampleSize}

\item{seed}{seed}

\item{wd}{working directory. The results will be saved here}

\item{total_repetitions}{total number of repetitions}
}
\description{
Simulation study 4
}
\examples{

}
\author{
Jannik Orzek
}
