% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation3.R
\name{simulation3}
\alias{simulation3}
\title{Simulation study 3}
\usage{
simulation3(sampleSize, seed, wd, total_repetitions)
}
\arguments{
\item{sampleSize}{sampleSize}

\item{seed}{seed}

\item{wd}{working directory. The results will be saved here}

\item{total_repetitions}{total number of repetitions}
}
\description{
Simulation study 3
}
\examples{

}
\author{
Jannik Orzek
}
